% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_109.R
\name{mdl_A_109}
\alias{mdl_A_109}
\title{stand (stem), age (mature)}
\usage{
mdl_A_109(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}
}
\value{
stand (stem), age (mature)
}
\description{
stand (stem), age (mature)
}
