% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_111.R
\name{mdl_A_111}
\alias{mdl_A_111}
\title{stand (stem), age (mature)}
\usage{
mdl_A_111(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), number of trees}

\item{p04}{stand (stem), diameter (breast height)}

\item{p05}{stand (stem), height (100 years old)}
}
\value{
stand (stem), age (mature)
}
\description{
stand (stem), age (mature)
}
