% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_24.R
\name{mdl_A_24}
\alias{mdl_A_24}
\title{single tree (stem), age (breast height)}
\usage{
mdl_A_24(p01, p02)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{stand (area), bonitet}
}
\value{
single tree (stem), age (breast height)
}
\description{
single tree (stem), age (breast height)
}
