% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0213.R
\name{mdl_BM_213}
\alias{mdl_BM_213}
\title{single tree (stem), biomass (above ground)}
\usage{
mdl_BM_213(p01, p02)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}
}
\value{
single tree (stem), biomass (above ground)
}
\description{
single tree (stem), biomass (above ground)
}
