% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_10.R
\name{mdl_D_10}
\alias{mdl_D_10}
\title{Stem D}
\usage{
mdl_D_10(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem A}

\item{p04}{Stem D}

\item{p05}{Area OHOR}
}
\value{
Stem D
}
\description{
Stem D
}
