% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_102.R
\name{mdl_D_102}
\alias{mdl_D_102}
\title{Stem D}
\usage{
mdl_D_102(p01, p02, p03, p04, p05, p06, p07, p08)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Stem D}

\item{p05}{Stem H}

\item{p06}{Stem A}

\item{p07}{Stem A}

\item{p08}{Stem G}
}
\value{
Stem D
}
\description{
Stem D
}
