% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_106.R
\name{mdl_D_106}
\alias{mdl_D_106}
\title{Stem D}
\usage{
mdl_D_106(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem D}

\item{p04}{Stem G}

\item{p05}{Stem G}
}
\value{
Stem D
}
\description{
Stem D
}
