% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_110.R
\name{mdl_D_110}
\alias{mdl_D_110}
\title{stand (stem), diameter (mature)}
\usage{
mdl_D_110(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}
}
\value{
stand (stem), diameter (mature)
}
\description{
stand (stem), diameter (mature)
}
