% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_136.R
\name{mdl_D_136}
\alias{mdl_D_136}
\title{single tree (stem), diameter (breast height)}
\usage{
mdl_D_136(p01, p02, p03, p04, p05, p06, p07)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (crown), height}

\item{p04}{stand (stem), basal area (breast height)}

\item{p05}{stand (larger trees), basal area (breast height)}

\item{p06}{stand (stem), height (dominant)}

\item{p07}{stand (stem), height (100 years old)}
}
\value{
single tree (stem), diameter (breast height)
}
\description{
single tree (stem), diameter (breast height)
}
