% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0150.R
\name{mdl_D_150}
\alias{mdl_D_150}
\title{single tree (crown), diameter}
\usage{
mdl_D_150(p01, p02, p03)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}
}
\value{
single tree (crown), diameter
}
\description{
single tree (crown), diameter
}
