% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_152.R
\name{mdl_D_152}
\alias{mdl_D_152}
\title{single tree (crown), diameter}
\usage{
mdl_D_152(p01, p02)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}
}
\value{
single tree (crown), diameter
}
\description{
single tree (crown), diameter
}
