% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_156.R
\name{mdl_D_156}
\alias{mdl_D_156}
\title{single tree (stem), diameter (random point)}
\usage{
mdl_D_156(p01, p02, p03, p04)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}

\item{p04}{single tree (stem), height (random point)}
}
\value{
single tree (stem), diameter (random point)
}
\description{
single tree (stem), diameter (random point)
}
