% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0183.R
\name{mdl_D_183}
\alias{mdl_D_183}
\title{stand (stem), diameter (breast height)}
\usage{
mdl_D_183(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), diameter (breast height)}

\item{p03}{stand (stem), age}

\item{p04}{stand (stem), age (random point)}
}
\value{
stand (stem), diameter (breast height)
}
\description{
stand (stem), diameter (breast height)
}
