% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_190.R
\name{mdl_D_190}
\alias{mdl_D_190}
\title{Stem D}
\usage{
mdl_D_190(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Stem G}
}
\value{
Stem D
}
\description{
Stem D
}
