% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_2.R
\name{mdl_D_2}
\alias{mdl_D_2}
\title{Stem D}
\usage{
mdl_D_2(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Stem H}
}
\value{
Stem D
}
\description{
Stem D
}
