% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0020.R
\name{mdl_D_20}
\alias{mdl_D_20}
\title{single tree (stem), diameter (breast height)}
\usage{
mdl_D_20(p01, p02)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (base)}
}
\value{
single tree (stem), diameter (breast height)
}
\description{
single tree (stem), diameter (breast height)
}
