% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_23.R
\name{mdl_D_23}
\alias{mdl_D_23}
\title{single tree (stem), diameter (base)}
\usage{
mdl_D_23(p01, p02)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}
}
\value{
single tree (stem), diameter (base)
}
\description{
single tree (stem), diameter (base)
}
