% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0034.R
\name{mdl_D_34}
\alias{mdl_D_34}
\title{stand (stem), diameter (breast height)}
\usage{
mdl_D_34(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}
}
\value{
stand (stem), diameter (breast height)
}
\description{
stand (stem), diameter (breast height)
}
