% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0043.R
\name{mdl_D_43}
\alias{mdl_D_43}
\title{species level (stem), diameter (breast height)}
\usage{
mdl_D_43(p01, p02)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}
}
\value{
species level (stem), diameter (breast height)
}
\description{
species level (stem), diameter (breast height)
}
