% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0045.R
\name{mdl_D_45}
\alias{mdl_D_45}
\title{stand (stem), diameter (breast height)}
\usage{
mdl_D_45(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), diameter (breast height)}
}
\value{
stand (stem), diameter (breast height)
}
\description{
stand (stem), diameter (breast height)
}
