% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_46.R
\name{mdl_D_46}
\alias{mdl_D_46}
\title{Stem D}
\usage{
mdl_D_46(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem D}

\item{p04}{Stem A}

\item{p05}{Stem A}
}
\value{
Stem D
}
\description{
Stem D
}
