% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_47.R
\name{mdl_D_47}
\alias{mdl_D_47}
\title{stand (stem), diameter (100 years old)}
\usage{
mdl_D_47(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}

\item{p03}{stand (stem), density}
}
\value{
stand (stem), diameter (100 years old)
}
\description{
stand (stem), diameter (100 years old)
}
