% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_8.R
\name{mdl_D_8}
\alias{mdl_D_8}
\title{single tree (bark), diameter ()}
\usage{
mdl_D_8(p01, p02, p03)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), height}

\item{p03}{single tree (stem), height (random point)}
}
\value{
single tree (bark), diameter ()
}
\description{
single tree (bark), diameter ()
}
