% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_159.R
\name{mdl_F_159}
\alias{mdl_F_159}
\title{single tree (stem), form}
\usage{
mdl_F_159(p01, p02, p03)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}
}
\value{
single tree (stem), form
}
\description{
single tree (stem), form
}
