% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0162.R
\name{mdl_F_162}
\alias{mdl_F_162}
\title{single tree (commercial wood), form}
\usage{
mdl_F_162(p01, p02, p03)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}
}
\value{
single tree (commercial wood), form
}
\description{
single tree (commercial wood), form
}
