% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_186.R
\name{mdl_F_186}
\alias{mdl_F_186}
\title{stand (stem), form}
\usage{
mdl_F_186(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}
}
\value{
stand (stem), form
}
\description{
stand (stem), form
}
