% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0120.R
\name{mdl_G_120}
\alias{mdl_G_120}
\title{stand (stem), basal area (breast height)}
\usage{
mdl_G_120(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}

\item{p03}{stand (stem), height (100 years old)}
}
\value{
stand (stem), basal area (breast height)
}
\description{
stand (stem), basal area (breast height)
}
