% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_141.R
\name{mdl_G_141}
\alias{mdl_G_141}
\title{species level (stem), basal area (breast height)}
\usage{
mdl_G_141(p01, p02, p03)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{species level (stem), height (20 years old)}
}
\value{
species level (stem), basal area (breast height)
}
\description{
species level (stem), basal area (breast height)
}
