% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_167.R
\name{mdl_G_167}
\alias{mdl_G_167}
\title{Stem G}
\usage{
mdl_G_167(p01, p02, p03, p04, p05, p06, p07)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem G}

\item{p03}{Stem A}

\item{p04}{Stem A}

\item{p05}{Stem N}

\item{p06}{Stem N}

\item{p07}{Stem H}
}
\value{
Stem G
}
\description{
Stem G
}
