% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_177.R
\name{mdl_G_177}
\alias{mdl_G_177}
\title{species level (stem), basal area (breast height)}
\usage{
mdl_G_177(p01, p02)
}
\arguments{
\item{p01}{species level (stem), species ()}

\item{p02}{species level (stem), height ()}
}
\value{
species level (stem), basal area (breast height)
}
\description{
species level (stem), basal area (breast height)
}
