% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_178.R
\name{mdl_G_178}
\alias{mdl_G_178}
\title{Stem G}
\usage{
mdl_G_178(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem G}

\item{p03}{Stem N}

\item{p04}{Stem N}

\item{p05}{Stem H}

\item{p06}{Stem H}
}
\value{
Stem G
}
\description{
Stem G
}
