% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_191.R
\name{mdl_G_191}
\alias{mdl_G_191}
\title{Stem G}
\usage{
mdl_G_191(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Stem G}
}
\value{
Stem G
}
\description{
Stem G
}
