% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0036.R
\name{mdl_G_36}
\alias{mdl_G_36}
\title{species level (stem), basal area (breast height)}
\usage{
mdl_G_36(p01, p02, p03)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{stand (stem), height (100 years old)}
}
\value{
species level (stem), basal area (breast height)
}
\description{
species level (stem), basal area (breast height)
}
