% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0073.R
\name{mdl_G_73}
\alias{mdl_G_73}
\title{stand (stem), basal area (breast height)}
\usage{
mdl_G_73(p01, p02, p03, p04, p05, p06, p07, p08, p09)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{species level (stem), basal area (breast height)}

\item{p03}{stand (stem), basal area (breast height)}

\item{p04}{stand (stem), number of trees}

\item{p05}{stand (stem), age}

\item{p06}{stand (stem), height (100 years old)}

\item{p07}{stand (area), latitude}

\item{p08}{stand (area), altitude}

\item{p09}{stand (stem), thinning exists}
}
\value{
stand (stem), basal area (breast height)
}
\description{
stand (stem), basal area (breast height)
}
