% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_88.R
\name{mdl_G_88}
\alias{mdl_G_88}
\title{stand (stem), basal area (breast height)}
\usage{
mdl_G_88(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}
}
\value{
stand (stem), basal area (breast height)
}
\description{
stand (stem), basal area (breast height)
}
