% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_1.R
\name{mdl_H_1}
\alias{mdl_H_1}
\title{single tree (stem), height}
\usage{
mdl_H_1(p01, p02)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}
}
\value{
single tree (stem), height
}
\description{
single tree (stem), height
}
