% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_107.R
\name{mdl_H_107}
\alias{mdl_H_107}
\title{Crwn H}
\usage{
mdl_H_107(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem G}

\item{p03}{Stem H}

\item{p04}{Stem D}

\item{p05}{Stem H}
}
\value{
Crwn H
}
\description{
Crwn H
}
