% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0170.R
\name{mdl_H_170}
\alias{mdl_H_170}
\title{stand (stem), height}
\usage{
mdl_H_170(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (dominant)}
}
\value{
stand (stem), height
}
\description{
stand (stem), height
}
