% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_174.R
\name{mdl_H_174}
\alias{mdl_H_174}
\title{stand (stem), height (dominant)}
\usage{
mdl_H_174(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}

\item{p03}{stand (stem), age}
}
\value{
stand (stem), height (dominant)
}
\description{
stand (stem), height (dominant)
}
