% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_205.R
\name{mdl_H_205}
\alias{mdl_H_205}
\title{Stem H}
\usage{
mdl_H_205(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem H}

\item{p03}{Stem D}

\item{p04}{Stem D}
}
\value{
Stem H
}
\description{
Stem H
}
