% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_28.R
\name{mdl_H_28}
\alias{mdl_H_28}
\title{stand (stem), height}
\usage{
mdl_H_28(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), height (100 years old)}
}
\value{
stand (stem), height
}
\description{
stand (stem), height
}
