% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0029.R
\name{mdl_H_29}
\alias{mdl_H_29}
\title{stand (stem), height}
\usage{
mdl_H_29(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (50 years old)}

\item{p03}{stand (stem), age}
}
\value{
stand (stem), height
}
\description{
stand (stem), height
}
