% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_44.R
\name{mdl_H_44}
\alias{mdl_H_44}
\title{species level (stem), height}
\usage{
mdl_H_44(p01, p02)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}
}
\value{
species level (stem), height
}
\description{
species level (stem), height
}
