% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_7.R
\name{mdl_H_7}
\alias{mdl_H_7}
\title{Stem H}
\usage{
mdl_H_7(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem H}

\item{p04}{Stem A}
}
\value{
Stem H
}
\description{
Stem H
}
