% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_80.R
\name{mdl_H_80}
\alias{mdl_H_80}
\title{Stem H}
\usage{
mdl_H_80(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem A}

\item{p04}{Stem H}
}
\value{
Stem H
}
\description{
Stem H
}
