% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_82.R
\name{mdl_H_82}
\alias{mdl_H_82}
\title{Stem H}
\usage{
mdl_H_82(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem D}

\item{p04}{Stem H}
}
\value{
Stem H
}
\description{
Stem H
}
