% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_87.R
\name{mdl_H_87}
\alias{mdl_H_87}
\title{stand (stem), height (dominant)}
\usage{
mdl_H_87(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}
}
\value{
stand (stem), height (dominant)
}
\description{
stand (stem), height (dominant)
}
