% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_9.R
\name{mdl_H_9}
\alias{mdl_H_9}
\title{Stem H}
\usage{
mdl_H_9(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem A}

\item{p04}{Stem H}

\item{p05}{Area Cult}

\item{p06}{Area OHOR}
}
\value{
Stem H
}
\description{
Stem H
}
