% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_hdif_ak.R
\name{mdl_hdif_ak}
\alias{mdl_hdif_ak}
\title{Stand level height growth model}
\usage{
mdl_hdif_ak(
  Species,
  Current_Age,
  Predict_Age,
  Current_Height,
  Cultivated,
  OHOR
)
}
\arguments{
\item{Species}{tree species}

\item{Current_Age}{current age}

\item{Predict_Age}{age of prediction}

\item{Current_Height}{current height}

\item{Cultivated}{is cultivated}

\item{OHOR}{O-horisont thickness (cm)}
}
\description{
Stand level height growth model
}
\examples{
mdl_hdif_ak("MA", 100, 110, 30, 1, 1)
}
\references{
Kiviste, A., 1999. Eesti puistute kasvumudelitest.
}
