% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_100.R
\name{mdl_HF_100}
\alias{mdl_HF_100}
\title{Stem HF}
\usage{
mdl_HF_100(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem HF}

\item{p03}{Stem H}

\item{p04}{Stem D}
}
\value{
Stem HF
}
\description{
Stem HF
}
