% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_135.R
\name{mdl_HF_135}
\alias{mdl_HF_135}
\title{Stem HF}
\usage{
mdl_HF_135(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Crwn H}

\item{p05}{Stem H}

\item{p06}{Stem G}
}
\value{
Stem HF
}
\description{
Stem HF
}
