% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_154.R
\name{mdl_HF_154}
\alias{mdl_HF_154}
\title{single tree (stem), formheight}
\usage{
mdl_HF_154(p01, p02, p03)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), height}

\item{p03}{single tree (stem), diameter (breast height)}
}
\value{
single tree (stem), formheight
}
\description{
single tree (stem), formheight
}
