% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0041.R
\name{mdl_HF_41}
\alias{mdl_HF_41}
\title{stand (stem), formheight}
\usage{
mdl_HF_41(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}
}
\value{
stand (stem), formheight
}
\description{
stand (stem), formheight
}
