% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0098.R
\name{mdl_HF_98}
\alias{mdl_HF_98}
\title{stand (stem), formheight}
\usage{
mdl_HF_98(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}

\item{p03}{stand (stem), basal area (breast height)}
}
\value{
stand (stem), formheight
}
\description{
stand (stem), formheight
}
