% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_99.R
\name{mdl_HF_99}
\alias{mdl_HF_99}
\title{Stem HF}
\usage{
mdl_HF_99(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem H}

\item{p03}{Stem D}

\item{p04}{Stem D}
}
\value{
Stem HF
}
\description{
Stem HF
}
