% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_115.R
\name{mdl_L_115}
\alias{mdl_L_115}
\title{stand (stem), distance between}
\usage{
mdl_L_115(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}

\item{p03}{stand (stem), diameter (breast height)}
}
\value{
stand (stem), distance between
}
\description{
stand (stem), distance between
}
