% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_127.R
\name{mdl_L_127}
\alias{mdl_L_127}
\title{stand (stem), distance between}
\usage{
mdl_L_127(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), diameter (breast height)}
}
\value{
stand (stem), distance between
}
\description{
stand (stem), distance between
}
