% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0193.R
\name{mdl_L_193}
\alias{mdl_L_193}
\title{stand (area), distance between ()}
\usage{
mdl_L_193(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species ()}

\item{p02}{stand (stem), diameter (breast height)}

\item{p03}{stand (stem), height (100 years old)}
}
\value{
stand (area), distance between ()
}
\description{
stand (area), distance between ()
}
