% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0037.R
\name{mdl_L_37}
\alias{mdl_L_37}
\title{stand (stem), distance between}
\usage{
mdl_L_37(p01, p02, p03, p04)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), diameter (breast height)}

\item{p03}{stand (stem), height (100 years old)}

\item{p04}{stand (area), humus layer}
}
\value{
stand (stem), distance between
}
\description{
stand (stem), distance between
}
