% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_105.R
\name{mdl_N_105}
\alias{mdl_N_105}
\title{stand (stem), number of trees}
\usage{
mdl_N_105(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), diameter (breast height)}

\item{p03}{stand (stem), height}
}
\value{
stand (stem), number of trees
}
\description{
stand (stem), number of trees
}
