% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0131.R
\name{mdl_N_131}
\alias{mdl_N_131}
\title{stand (stem), number of trees}
\usage{
mdl_N_131(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}
}
\value{
stand (stem), number of trees
}
\description{
stand (stem), number of trees
}
