% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_139.R
\name{mdl_N_139}
\alias{mdl_N_139}
\title{species level (stem), number of trees}
\usage{
mdl_N_139(p01, p02, p03, p04)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{species level (stem), basal area (breast height)}

\item{p04}{species level (stem), height (20 years old)}
}
\value{
species level (stem), number of trees
}
\description{
species level (stem), number of trees
}
