% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_140.R
\name{mdl_N_140}
\alias{mdl_N_140}
\title{species level (stem), number of trees}
\usage{
mdl_N_140(p01, p02)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), diameter (breast height)}
}
\value{
species level (stem), number of trees
}
\description{
species level (stem), number of trees
}
