% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_168.R
\name{mdl_N_168}
\alias{mdl_N_168}
\title{Stem N}
\usage{
mdl_N_168(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem N}

\item{p03}{Stem H}

\item{p04}{Stem H}
}
\value{
Stem N
}
\description{
Stem N
}
