% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_54.R
\name{mdl_N_54}
\alias{mdl_N_54}
\title{stand (stem), number of trees}
\usage{
mdl_N_54(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), diameter}

\item{p03}{stand (stem), diameter (base)}

\item{p04}{stand (stem), number of trees (base)}
}
\value{
stand (stem), number of trees
}
\description{
stand (stem), number of trees
}
