% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0056.R
\name{mdl_N_56}
\alias{mdl_N_56}
\title{stand (stem), number of trees (base)}
\usage{
mdl_N_56(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species ()}

\item{p02}{stand (stem), number of trees (random point)}

\item{p03}{stand (stem), diameter (breast height)}

\item{p04}{stand (stem), diameter (breast height)}
}
\value{
stand (stem), number of trees (base)
}
\description{
stand (stem), number of trees (base)
}
