% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_12.R
\name{mdl_Price_12}
\alias{mdl_Price_12}
\title{Stem Price}
\usage{
mdl_Price_12(p01, p02, p03)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}
}
\value{
Stem Price
}
\description{
Stem Price
}
