% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0081.R
\name{mdl_Prob_81}
\alias{mdl_Prob_81}
\title{single tree (stem), probability}
\usage{
mdl_Prob_81(p01, p02, p03, p04, p05, p06, p07, p08)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{stand (stem), height (100 years old)}

\item{p04}{stand (larger trees), basal area (breast height)}

\item{p05}{stand (stem), basal area (breast height)}

\item{p06}{species level (stem), basal area (breast height)}

\item{p07}{single tree (stem), age}

\item{p08}{single tree (stem), age (random point)}
}
\value{
single tree (stem), probability
}
\description{
single tree (stem), probability
}
