% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0146.R
\name{mdl_T_146}
\alias{mdl_T_146}
\title{stand (stem), density}
\usage{
mdl_T_146(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height (100 years old)}

\item{p03}{stand (stem), age}
}
\value{
stand (stem), density
}
\description{
stand (stem), density
}
