% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_11.R
\name{mdl_V_11}
\alias{mdl_V_11}
\title{Stem V}
\usage{
mdl_V_11(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem A}

\item{p04}{Stem V}

\item{p05}{Area OHOR}

\item{p06}{Area Cult}
}
\value{
Stem V
}
\description{
Stem V
}
