% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0013.R
\name{mdl_V_13}
\alias{mdl_V_13}
\title{species level (stem), volume}
\usage{
mdl_V_13(p01, p02, p03, p04)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{stand (stem), density}

\item{p04}{stand (stem), height (100 years old)}
}
\value{
species level (stem), volume
}
\description{
species level (stem), volume
}
