% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_14.R
\name{mdl_V_14}
\alias{mdl_V_14}
\title{Stem V}
\usage{
mdl_V_14(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem T}

\item{p04}{Stem H}
}
\value{
Stem V
}
\description{
Stem V
}
