% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_144.R
\name{mdl_V_144}
\alias{mdl_V_144}
\title{species level (stem), volume}
\usage{
mdl_V_144(p01, p02, p03, p04)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{stand (stem), height (100 years old)}

\item{p04}{stand (stem), density}
}
\value{
species level (stem), volume
}
\description{
species level (stem), volume
}
