% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_147.R
\name{mdl_V_147}
\alias{mdl_V_147}
\title{Stem V}
\usage{
mdl_V_147(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem G}

\item{p03}{Stem A}

\item{p04}{Stem Bon}
}
\value{
Stem V
}
\description{
Stem V
}
