% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0172.R
\name{mdl_V_172}
\alias{mdl_V_172}
\title{single tree (bark), volume}
\usage{
mdl_V_172(p01, p02)
}
\arguments{
\item{p01}{single tree (bark), species}

\item{p02}{single tree (stem), diameter (breast height)}
}
\value{
single tree (bark), volume
}
\description{
single tree (bark), volume
}
