% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_179.R
\name{mdl_V_179}
\alias{mdl_V_179}
\title{stand (stem), volume}
\usage{
mdl_V_179(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), bonitet}
}
\value{
stand (stem), volume
}
\description{
stand (stem), volume
}
