% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_184.R
\name{mdl_V_184}
\alias{mdl_V_184}
\title{Stem V}
\usage{
mdl_V_184(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem V}

\item{p03}{Stem A}

\item{p04}{Stem A}
}
\value{
Stem V
}
\description{
Stem V
}
