% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_187.R
\name{mdl_V_187}
\alias{mdl_V_187}
\title{Stem V}
\usage{
mdl_V_187(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Stem N}
}
\value{
Stem V
}
\description{
Stem V
}
