% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0021.R
\name{mdl_V_21}
\alias{mdl_V_21}
\title{species level (stem), volume}
\usage{
mdl_V_21(p01, p02, p03)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), height}

\item{p03}{species level (stem), basal area (breast height)}
}
\value{
species level (stem), volume
}
\description{
species level (stem), volume
}
