% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_3.R
\name{mdl_V_3}
\alias{mdl_V_3}
\title{species level (stem), volume}
\usage{
mdl_V_3(p01, p02)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), height ()}
}
\value{
species level (stem), volume
}
\description{
species level (stem), volume
}
