% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_40.R
\name{mdl_V_40}
\alias{mdl_V_40}
\title{stand (stem), volume}
\usage{
mdl_V_40(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}

\item{p03}{stand (stem), age}
}
\value{
stand (stem), volume
}
\description{
stand (stem), volume
}
