% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_42.R
\name{mdl_V_42}
\alias{mdl_V_42}
\title{species level (stem), volume}
\usage{
mdl_V_42(p01, p02, p03, p04, p05, p06, p07, p08, p09, p10, p11, p12)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), basal area (breast height)}

\item{p03}{stand (stem), basal area (breast height)}

\item{p04}{species level (stem), number of trees}

\item{p05}{stand (stem), number of trees}

\item{p06}{stand (area), vegetation layer exists}

\item{p07}{stand (area), thinning exists}

\item{p08}{stand (stem), height (100 years old)}

\item{p09}{species level (stem), diameter (breast height)}

\item{p10}{stand (area), altitude}

\item{p11}{stand (area), latitude}

\item{p12}{species level (stem), age}
}
\value{
species level (stem), volume
}
\description{
species level (stem), volume
}
