% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0053.R
\name{mdl_V_53}
\alias{mdl_V_53}
\title{stand (stem), volume}
\usage{
mdl_V_53(p01, p02)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}
}
\value{
stand (stem), volume
}
\description{
stand (stem), volume
}
