% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_67.R
\name{mdl_V_67}
\alias{mdl_V_67}
\title{single tree (stem), volume}
\usage{
mdl_V_67(p01, p02, p03)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}
}
\value{
single tree (stem), volume
}
\description{
single tree (stem), volume
}
