% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0096.R
\name{mdl_V_96}
\alias{mdl_V_96}
\title{stand (stem), volume}
\usage{
mdl_V_96(p01, p02, p03)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age (breast height)}

\item{p03}{stand (stem), volume}
}
\value{
stand (stem), volume
}
\description{
stand (stem), volume
}
