% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_111.R
\name{mdl_A_111}
\alias{mdl_A_111}
\title{Stem A}
\usage{
mdl_A_111(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem N}

\item{p04}{Stem D}

\item{p05}{Stem H}
}
\value{
Stem A
}
\description{
Stem A
}
