% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_10.R
\name{mdl_D_10}
\alias{mdl_D_10}
\title{stand (stem), diameter (breast height)}
\usage{
mdl_D_10(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), age (random point)}

\item{p04}{stand (stem), diameter (breast height)}

\item{p05}{stand (area), humus layer}
}
\value{
stand (stem), diameter (breast height)
}
\description{
stand (stem), diameter (breast height)
}
