% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_102.R
\name{mdl_D_102}
\alias{mdl_D_102}
\title{single tree (stem), diameter (breast height)}
\usage{
mdl_D_102(p01, p02, p03, p04, p05, p06, p07, p08)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}

\item{p04}{stand (stem), diameter (breast height)}

\item{p05}{stand (stem), height}

\item{p06}{single tree (stem), age}

\item{p07}{stand (stem), age}

\item{p08}{stand (stem), basal area (breast height)}
}
\value{
single tree (stem), diameter (breast height)
}
\description{
single tree (stem), diameter (breast height)
}
