% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_106.R
\name{mdl_D_106}
\alias{mdl_D_106}
\title{single tree (stem), diameter (breast height)}
\usage{
mdl_D_106(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{single tree (stem), diameter (breast height)}

\item{p04}{species level (stem), basal area (breast height)}

\item{p05}{stand (stem), basal area (breast height)}
}
\value{
single tree (stem), diameter (breast height)
}
\description{
single tree (stem), diameter (breast height)
}
