% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_136.R
\name{mdl_D_136}
\alias{mdl_D_136}
\title{Stem D}
\usage{
mdl_D_136(p01, p02, p03, p04, p05, p06, p07)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Crwn H}

\item{p04}{Stem G}

\item{p05}{LarTr G}

\item{p06}{Stem H}

\item{p07}{Stem H}
}
\value{
Stem D
}
\description{
Stem D
}
