% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_30.R
\name{mdl_D_30}
\alias{mdl_D_30}
\title{Stem D}
\usage{
mdl_D_30(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem H}

\item{p03}{Stem A}

\item{p04}{Stem H}
}
\value{
Stem D
}
\description{
Stem D
}
