% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0167.R
\name{mdl_G_167}
\alias{mdl_G_167}
\title{stand (stem), basal area (breast height)}
\usage{
mdl_G_167(p01, p02, p03, p04, p05, p06, p07)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), basal area (breast height)}

\item{p03}{stand (stem), age}

\item{p04}{stand (stem), age (random point)}

\item{p05}{stand (stem), number of trees}

\item{p06}{stand (stem), number of trees}

\item{p07}{stand (stem), height (100 years old)}
}
\value{
stand (stem), basal area (breast height)
}
\description{
stand (stem), basal area (breast height)
}
