% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0178.R
\name{mdl_G_178}
\alias{mdl_G_178}
\title{stand (stem), basal area (breast height)}
\usage{
mdl_G_178(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), basal area (breast height)}

\item{p03}{stand (stem), number of trees}

\item{p04}{stand (stem), number of trees}

\item{p05}{stand (stem), height}

\item{p06}{stand (stem), height}
}
\value{
stand (stem), basal area (breast height)
}
\description{
stand (stem), basal area (breast height)
}
