% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_191.R
\name{mdl_G_191}
\alias{mdl_G_191}
\title{species level (stem), basal area (breast height)}
\usage{
mdl_G_191(p01, p02, p03, p04)
}
\arguments{
\item{p01}{species level (stem), species ()}

\item{p02}{species level (stem), diameter (breast height)}

\item{p03}{stand (stem), height (100 years old)}

\item{p04}{species level (stem), basal area (breast height)}
}
\value{
species level (stem), basal area (breast height)
}
\description{
species level (stem), basal area (breast height)
}
