% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_73.R
\name{mdl_G_73}
\alias{mdl_G_73}
\title{Stem G}
\usage{
mdl_G_73(p01, p02, p03, p04, p05, p06, p07, p08, p09)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem G}

\item{p03}{Stem G}

\item{p04}{Stem N}

\item{p05}{Stem A}

\item{p06}{Stem H}

\item{p07}{Area Lat}

\item{p08}{Area Alt}

\item{p09}{Stem Thin}
}
\value{
Stem G
}
\description{
Stem G
}
