% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_h100.R
\name{mdl_h100}
\alias{mdl_h100}
\title{Site index model}
\usage{
mdl_h100(Species, Stand_Age, Stand_Height, Base_Age = 100)
}
\arguments{
\item{Species}{tree species}

\item{Stand_Age}{main species age}

\item{Stand_Height}{main species average height}

\item{Base_Age}{Age of site index height}
}
\description{
Site index model
}
\examples{
mdl_h100("MA", 100, 20, seq(60, 140, by=20))

}
\author{
Allan Sims
}
