% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_107.R
\name{mdl_H_107}
\alias{mdl_H_107}
\title{single tree (crown), height}
\usage{
mdl_H_107(p01, p02, p03, p04, p05)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{stand (stem), basal area (breast height)}

\item{p03}{stand (stem), height (dominant)}

\item{p04}{single tree (stem), diameter (breast height)}

\item{p05}{single tree (stem), height}
}
\value{
single tree (crown), height
}
\description{
single tree (crown), height
}
