% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_182.R
\name{mdl_H_182}
\alias{mdl_H_182}
\title{stand (stem), height}
\usage{
mdl_H_182(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}

\item{p03}{stand (stem), age}

\item{p04}{stand (stem), age (random point)}
}
\value{
stand (stem), height
}
\description{
stand (stem), height
}
