% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0038.R
\name{mdl_H_38}
\alias{mdl_H_38}
\title{single tree (stem), height}
\usage{
mdl_H_38(p01, p02, p03, p04)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{species level (stem), diameter (breast height)}

\item{p04}{species level (stem), height}
}
\value{
single tree (stem), height
}
\description{
single tree (stem), height
}
