% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_7.R
\name{mdl_H_7}
\alias{mdl_H_7}
\title{species level (stem), height ()}
\usage{
mdl_H_7(p01, p02, p03, p04)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), age}

\item{p03}{species level (stem), height}

\item{p04}{species level (stem), age (random point)}
}
\value{
species level (stem), height ()
}
\description{
species level (stem), height ()
}
