% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0080.R
\name{mdl_H_80}
\alias{mdl_H_80}
\title{stand (stem), height}
\usage{
mdl_H_80(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age (breast height)}

\item{p03}{stand (stem), age (breast height)}

\item{p04}{stand (stem), height}
}
\value{
stand (stem), height
}
\description{
stand (stem), height
}
