% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_9.R
\name{mdl_H_9}
\alias{mdl_H_9}
\title{stand (stem), height}
\usage{
mdl_H_9(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), age (random point)}

\item{p04}{stand (stem), height ()}

\item{p05}{stand (area), cultivated}

\item{p06}{stand (area), humus layer}
}
\value{
stand (stem), height
}
\description{
stand (stem), height
}
