% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_135.R
\name{mdl_HF_135}
\alias{mdl_HF_135}
\title{single tree (stem), formheight}
\usage{
mdl_HF_135(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), diameter (breast height)}

\item{p03}{single tree (stem), height}

\item{p04}{single tree (crown), height}

\item{p05}{stand (stem), height (dominant)}

\item{p06}{stand (stem), basal area (breast height)}
}
\value{
single tree (stem), formheight
}
\description{
single tree (stem), formheight
}
