% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0099.R
\name{mdl_HF_99}
\alias{mdl_HF_99}
\title{single tree (stem), formheight}
\usage{
mdl_HF_99(p01, p02, p03, p04)
}
\arguments{
\item{p01}{single tree (stem), species}

\item{p02}{single tree (stem), height}

\item{p03}{single tree (stem), diameter (breast height)}

\item{p04}{single tree (stem), diameter (6 meters)}
}
\value{
single tree (stem), formheight
}
\description{
single tree (stem), formheight
}
