% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_37.R
\name{mdl_L_37}
\alias{mdl_L_37}
\title{Stem L}
\usage{
mdl_L_37(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{Area OHOR}
}
\value{
Stem L
}
\description{
Stem L
}
