% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_139.R
\name{mdl_N_139}
\alias{mdl_N_139}
\title{Stem N}
\usage{
mdl_N_139(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem G}

\item{p04}{Stem H}
}
\value{
Stem N
}
\description{
Stem N
}
