% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_168.R
\name{mdl_N_168}
\alias{mdl_N_168}
\title{stand (stem), number of trees ()}
\usage{
mdl_N_168(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), number of trees}

\item{p03}{stand (stem), height}

\item{p04}{stand (stem), height}
}
\value{
stand (stem), number of trees ()
}
\description{
stand (stem), number of trees ()
}
