% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_12.R
\name{mdl_Price_12}
\alias{mdl_Price_12}
\title{species level (stem), price}
\usage{
mdl_Price_12(p01, p02, p03)
}
\arguments{
\item{p01}{species level (stem), species}

\item{p02}{species level (stem), diameter (breast height)}

\item{p03}{species level (stem), height}
}
\value{
species level (stem), price
}
\description{
species level (stem), price
}
