% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_81.R
\name{mdl_Prob_81}
\alias{mdl_Prob_81}
\title{Stem Prob}
\usage{
mdl_Prob_81(p01, p02, p03, p04, p05, p06, p07, p08)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem H}

\item{p04}{LarTr G}

\item{p05}{Stem G}

\item{p06}{Stem G}

\item{p07}{Stem A}

\item{p08}{Stem A}
}
\value{
Stem Prob
}
\description{
Stem Prob
}
