% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_11.R
\name{mdl_V_11}
\alias{mdl_V_11}
\title{stand (stem), volume ()}
\usage{
mdl_V_11(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), age}

\item{p03}{stand (stem), age (random point)}

\item{p04}{stand (stem), volume}

\item{p05}{stand (area), humus layer}

\item{p06}{stand (area), cultivated}
}
\value{
stand (stem), volume ()
}
\description{
stand (stem), volume ()
}
