% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_144.R
\name{mdl_V_144}
\alias{mdl_V_144}
\title{Stem V}
\usage{
mdl_V_144(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem A}

\item{p03}{Stem H}

\item{p04}{Stem T}
}
\value{
Stem V
}
\description{
Stem V
}
