% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_184.R
\name{mdl_V_184}
\alias{mdl_V_184}
\title{stand (stem), volume ()}
\usage{
mdl_V_184(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species ()}

\item{p02}{stand (stem), volume ()}

\item{p03}{stand (stem), age ()}

\item{p04}{stand (stem), age (random point)}
}
\value{
stand (stem), volume ()
}
\description{
stand (stem), volume ()
}
