% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_0187.R
\name{mdl_V_187}
\alias{mdl_V_187}
\title{stand (stem), volume ()}
\usage{
mdl_V_187(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), diameter (breast height)}

\item{p03}{stand (stem), height ()}

\item{p04}{stand (stem), number of trees}
}
\value{
stand (stem), volume ()
}
\description{
stand (stem), volume ()
}
