% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_42.R
\name{mdl_V_42}
\alias{mdl_V_42}
\title{Stem V}
\usage{
mdl_V_42(p01, p02, p03, p04, p05, p06, p07, p08, p09, p10, p11, p12)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem G}

\item{p03}{Stem G}

\item{p04}{Stem N}

\item{p05}{Stem N}

\item{p06}{Area Veg}

\item{p07}{Area Thin}

\item{p08}{Stem H}

\item{p09}{Stem D}

\item{p10}{Area Alt}

\item{p11}{Area Lat}

\item{p12}{Stem A}
}
\value{
Stem V
}
\description{
Stem V
}
