% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_78.R
\name{mdl_V_78}
\alias{mdl_V_78}
\title{Stem V}
\usage{
mdl_V_78(p01, p02, p03, p04)
}
\arguments{
\item{p01}{Stem Spec}

\item{p02}{Stem D}

\item{p03}{Stem D}

\item{p04}{Stem H}
}
\value{
Stem V
}
\description{
Stem V
}
