% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_95.R
\name{mdl_V_95}
\alias{mdl_V_95}
\title{stand (stem), volume}
\usage{
mdl_V_95(p01, p02, p03, p04)
}
\arguments{
\item{p01}{stand (stem), species}

\item{p02}{stand (stem), height}

\item{p03}{stand (stem), basal area (breast height)}

\item{p04}{stand (stem), age}
}
\value{
stand (stem), volume
}
\description{
stand (stem), volume
}
